fix(read): 滚动 slop/音量键/滚轮收尾, 封面选图与平台失效推送区分
- 滚动模式: 补第二道欧氏 slop 判定, slop 双标志对齐原版; 修滚动抖动与偏移回灌
- 音量键翻页: 恒定启用, 补 Activity 层 dispatchKeyEvent/dispatchShortcut 桥接; 移除滚轮翻页设置项, 滚轮翻页彻底禁用
- 四行样式: 补翻列吸附, fling 限 ±1 列 + 90ms/inch 黏滞
- 书源详情: 修刷新封面无效与本地图片被静默拒; 封面选图持久落盘 + 编辑保存回传不回顶
- 编辑器: 修回车缩进丢失/空布局越界崩溃/长文本点击乱跳
- 平台能力: 新增 expect/actual isAndroidPlatform, 非 Android (桌面/iOS/鸿蒙) 书架 UP_BOOKSHELF 改走激活期 500ms 聚合重查 (桌面 Room UPDATE 失效推送不可靠), Android 维持原可见时兜底; AppShortcuts preemptive 与成员属性同名致快捷键 KeyDown 永不匹配, 改名 resolvedPreemptive 根治

+++ shared/src/sharedUiMain/kotlin/io/legado/app/ui/book/read/page/delegate/ScrollPageDelegateCompose.kt
@@ -505,11 +505,13 @@ class ScrollPageDelegateCompose(
             decaySpec = splineBasedDecay(LocalDensity.current)
         }
 
-        // 与 viewModel.scrollOffset 同步: 切章/重排重置归零时本 delegate 跟随
-        // (applyScrollDelta 已双写, 此处只消费外部重置, 同值写入不触发重绘)
+        // 与 viewModel.scrollOffset 同步: 只消费外部重置 (切章/重排归零), 不回灌自己的写入。
+        // 比较必须在 Int 空间做: 本 delegate 的偏移是 Float, 双写时 updateScrollOffset 截断成
+        // Int, 若按 Float 比较则平滑滚动的每一帧都"不等"→ 每帧把小数抹掉并回写, 与正在进行的
+        // 手势/惯性互相打架, 表现为滚动抖动/闪现几行
         LaunchedEffect(viewModel) {
             viewModel.scrollOffset.collect { offset ->
-                if (_currentOffset != offset.toFloat()) {
+                if (_currentOffset.toInt() != offset) {
                     _currentOffset = offset.toFloat()
                 }
             }
